Trim whitespace from the `current_todo` variable
(Line 735) The `current_todo` item did not match the same item inside of the `todo_list`, because the `current_todo` item included some extra whitespace at the end of the string, therefore the if statement at line 736 always returned false. The function `rstrip()` fixes this, since it strips the whitespace at the end of the string. Please note that this is a temporary fix, and it should be reviewed, as I have only tested it with my setup.