Had to undo commit 880cdc2
After this commit, PyOrgMode did not match any of my test org mode files with TODOs on them. I have not tested exactly *why* that is the case, however I know it's the culprit. I'm not exactly sure what purpose it's for, since the author of the commit stated that without this addition, the library was not able to properly list headings with dashes in them, however I did not encounter this with my test files (after adding a dash in one of the entries). I tested it by printing each node (Line 728) and seeing whether the headings can display dashes and running `base.extract_todo_list()[0].heading` to see if the heading printed out correctly, and it did (was not able to run the tests since I did not have the locale installed). However, this commit requries thorough review, as I have not tested each aspect of this program.